add ultils tests

+++ masr/ultils/pressor.py
@@ -1,8 +1,15 @@
 import zstandard as zstd
+import os
 
 
 def compress_file(input_file_path, output_file_path, compression_level=3):
     try:
+        # 检查输入文件是否存在
+        if not os.path.isfile(input_file_path):
+            raise FileNotFoundError(f"compress_file: \
+                                    No file found at specified path: \
+                                    {input_file_path}")
+
         with open(input_file_path, "rb") as input_file:
             with open(output_file_path, "wb") as output_file:
                 compressor = zstd.ZstdCompressor(level=compression_level)
@@ -15,6 +22,12 @@ def compress_file(input_file_path, output_file_path, compression_level=3):
 
 def decompress_file(input_file_path, output_file_path):
     try:
+        # 检查输入文件是否存在
+        if not os.path.isfile(input_file_path):
+            raise FileNotFoundError(f"decompress_file: \
+                                    No file found at specified path: \
+                                        {input_file_path}")
+
         with open(input_file_path, "rb") as input_file:
             with open(output_file_path, "wb") as output_file:
                 compressor = zstd.ZstdDecompressor()
